feat: integrate speech into text chat
